Implement Saving a Session

+++ ViewModels/SessionsFormViewModel.cs
@@ -8,7 +8,7 @@ namespace MentorHub.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; } 
         public string Time { get; set; }
-        public int Occupations { get; set; }
+        public byte Occupations { get; set; }
         public IEnumerable<Occupations> Profession { get; set; }
 
     }
